Inlined 'Array.tabulate' into 'DenseVector.tabulate'
This allows to avoid boxing in the argument and return type of 'f'.
You  might be thinking "Boxing? but it is specialized!". Turns out
'Array.tabulate' is not, therefore it would call a non-specialized
version of the 'Function1' trait, which would effectively box both
the index and the produced value.