Don't insert default value in OpenHashArray
If we are inserting a new element, and that element is the default, then
we don't do anything and don't increase the load factor.

NB: default value can still be inserted if it is inserted after a non-
default value that was inserted with the same key. We could remove the
entry in that case but that would require more work